inventory/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2
django-admin-autocomplete-filter==0.7.1
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
